Split Rest and Interval to 2 components
Add startButton component
Add TimerInput component as base component for Rest and Interval
Also, fixed few bugs
