+++ src/components/CountSpeed/CountSpeedReducers.js
@@ -1,4 +1,4 @@
-export const CountSpeed = (state = { countSpeed: 1 } ,action) => {
+export const CountSpeed = (state = { countSpeed: 1000 } ,action) => {
     switch (action.type){
         case "CHANGE_SPEED":
             return { countSpeed: action.payload}
