+++ src/components/CountToNumber/CountToNumber.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import {PlusNumber, MinusNumber} from './CountToNumberActions';
+import { PlusNumber, MinusNumber } from './CountToNumberActions';
 import NumberCounter from '../GeneralComponents/NumberCounter';
 
 class CountToNumber extends Component {
