+++ src/components/IntervalTimer/IntervalTimerActions.js
@@ -1,0 +1,23 @@
+export const IntervalMinPlus = () => {
+    return {
+        type: 'INTERVAL_MIN_PLUS'
+    };
+};
+
+export const IntervalMinMinus = () => {
+    return {
+        type: 'INTERVAL_MIN_MINUS'
+    };
+};
+
+export const IntervalSecPlus = () => {
+    return {
+        type: 'INTERVAL_SEC_PLUS'
+    };
+};
+
+export const IntervalSecMinus = () => {
+    return {
+        type: 'INTERVAL_SEC_MINUS'
+    };
+};
