+++ src/components/RestTimer/RestTimerActions.js
@@ -1,0 +1,23 @@
+export const RestMinPlus = () => {
+    return {
+        type: 'REST_MIN_PLUS'
+    };
+};
+
+export const RestMinMinus = () => {
+    return {
+        type: 'REST_MIN_MINUS'
+    };
+};
+
+export const RestSecPlus = () => {
+    return {
+        type: 'REST_SEC_PLUS'
+    };
+};
+
+export const RestSecMinus = () => {
+    return {
+        type: 'REST_SEC_MINUS'
+    };
+};
