--- src/components/TimerInput/TimerInputActions.js
@@ -1,50 +1,0 @@
-// ### INTERVAL ###
-export const IntervalMinPlus = () => {
-    return {
-        type: 'INTERVAL_MIN_PLUS'
-    };
-};
-
-export const IntervalMinMinus = () => {
-    return {
-        type: 'INTERVAL_MIN_MINUS'
-    };
-};
-
-export const IntervalSecPlus = () => {
-    return {
-        type: 'INTERVAL_SEC_PLUS'
-    };
-};
-
-export const IntervalSecMinus = () => {
-    return {
-        type: 'INTERVAL_SEC_MINUS'
-    };
-};
-
-// ### REST ###
-
-export const RestMinPlus = () => {
-    return {
-        type: 'REST_MIN_PLUS'
-    };
-};
-
-export const RestMinMinus = () => {
-    return {
-        type: 'REST_MIN_MINUS'
-    };
-};
-
-export const RestSecPlus = () => {
-    return {
-        type: 'REST_SEC_PLUS'
-    };
-};
-
-export const RestSecMinus = () => {
-    return {
-        type: 'REST_SEC_MINUS'
-    };
-};
